streamlining feature extraction scripts

+++ AcousticFeatureExtraction/Scripts/master.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 
+import os
 import extract
 import preProcess
 import subprocess
@@ -7,16 +8,18 @@ import limitsUpperLower
 
 def main(wavPath, speakerList, outputType, winSize=10, prune=True, needReaper=False, needAMS=False, needPLP=False):
 
-    #convert stereo audio to mono
-    #downsample to 16000 Hz
-    bashCommand = "cd ../../AudioData/{}; mkdir downsampled; ../../AcousticFeatureExtraction/Scripts/monoDown.sh; mkdir ../temp{}; mv downsampled/* ../temp{}; rm -r downsampled; cd ../../AcousticFeatureExtraction/Scripts".format(wavPath, wavPath, wavPath)
-    subprocess.run(bashCommand, shell=True)
+    if not os.path.isdir("../../AudioData/Gated{}".format(wavPath)):
 
-    #preProcess.py normalizes rms to the average for all files (time-consuming), and trims leading and trailing silences
-    preProcess.main(wavPath)
+        #convert stereo audio to mono
+        #downsample to 16000 Hz
+        bashCommand = "cd ../../AudioData/{}; mkdir downsampled; ../../AcousticFeatureExtraction/Scripts/monoDown.sh; mkdir ../temp{}; mv downsampled/* ../temp{}; rm -r downsampled; cd ../../AcousticFeatureExtraction/Scripts".format(wavPath, wavPath, wavPath)
+        subprocess.run(bashCommand, shell=True)
 
-    #Get rid of temp folder. Files now live in Gated{wavPath}
-    subprocess.run("rm -r ../../AudioData/temp{}".format(wavPath), shell=True)
+        #preProcess.py normalizes rms to the average for all files (time-consuming), and trims leading and trailing silences
+        preProcess.main(wavPath)
+
+        #Get rid of temp folder. Files now live in Gated{wavPath}
+        subprocess.run("rm -r ../../AudioData/temp{}".format(wavPath), shell=True)
 
     if needReaper == True:
         bashCommand = "mkdir ../ReaperTxtFiles/{}_{}ms_ReaperF0Results; cd ../../AudioData/Gated{}; ../../AcousticFeatureExtraction/Scripts/REAPER/reaper.sh; ../../AcousticFeatureExtraction/Scripts/REAPER/formatReaperOutputs.sh; mv *.p ../../AcousticFeatureExtraction/ReaperTxtFiles/{}_{}ms_ReaperF0Results/; rm *.f0; cd ../../AcousticFeatureExtraction/Scripts".format(wavPath, winSize, wavPath, wavPath, winSize)
